+++ src/renderer/core/music-sheet/common/default-sheet.ts
