--- src/renderer/core/music-sheet/internal/store.ts
@@ -1,5 +1,0 @@
-import Store from "@/common/store";
-import defaultSheet from "./default-sheet";
-
-export const musicSheetsStore = new Store<IMusic.IDBMusicSheetItem[]>([defaultSheet]);
-export const starredSheetsStore = new Store<IMedia.IMediaBase[]>([]);
