+++ src/renderer/pages/main-page/views/music-sheet-view/local-sheet/index.tsx
@@ -1,16 +1,16 @@
 import { useParams } from "react-router-dom";
-import { useMusicSheet } from "@/renderer/core/music-sheet/internal/sheets-method";
 import MusicSheetlikeView from "@/renderer/components/MusicSheetlikeView";
 import { RequestStateCode } from "@/common/constant";
+import MusicSheet from "@/renderer/core/music-sheet";
 
 export default function LocalSheet() {
   const { id } = useParams() ?? {};
-  const [musicSheet, loading] = useMusicSheet(id);
+  const [musicSheet, loading] = MusicSheet.frontend.useMusicSheet(id);
 
   return (
     <MusicSheetlikeView
       musicSheet={musicSheet}
-      state={loading ? RequestStateCode.LOADING : RequestStateCode.FINISHED}
+      state={loading}
       musicList={musicSheet?.musicList ?? []}
     ></MusicSheetlikeView>
   );
