+++ src/renderer/pages/main-page/views/music-sheet-view/remote-sheet/index.tsx
@@ -2,22 +2,17 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import usePluginSheetMusicList from "./hooks/usePluginSheetMusicList";
 import MusicSheetlikeView from "@/renderer/components/MusicSheetlikeView";
-import { starredSheetsStore } from "@/renderer/core/music-sheet";
 import { isSameMedia } from "@/common/media-util";
-import {
-  starMusicSheet,
-  unstarMusicSheet,
-} from "@/renderer/core/music-sheet/internal/sheets-method";
+
+import MusicSheet from "@/renderer/core/music-sheet";
 
 export default function RemoteSheet() {
   const { platform, id } = useParams() ?? {};
 
   const [state, sheetItem, musicList, getSheetDetail] = usePluginSheetMusicList(
-    {
-      ...(history.state?.usr?.sheetItem ?? {}),
-      platform,
-      id,
-    } as IMusic.IMusicSheetItem
+    platform,
+    id,
+    history.state?.usr?.sheetItem
   );
   return (
     <MusicSheetlikeView
@@ -37,7 +32,7 @@ interface IProps {
 }
 function RemoteSheetOptions(props: IProps) {
   const { sheetItem } = props;
-  const starredMusicSheets = starredSheetsStore.useValue();
+  const starredMusicSheets = MusicSheet.frontend.useAllStarredSheets();
 
   const isStarred = starredMusicSheets.find((item) =>
     isSameMedia(sheetItem, item)
@@ -49,7 +44,9 @@ function RemoteSheetOptions(props: IProps) {
         role="button"
         data-type="normalButton"
         onClick={() => {
-          isStarred ? unstarMusicSheet(sheetItem) : starMusicSheet(sheetItem);
+          isStarred
+            ? MusicSheet.frontend.unstarMusicSheet(sheetItem)
+            : MusicSheet.frontend.starMusicSheet(sheetItem);
         }}
       >
         {isStarred ? "取消收藏" : "收藏歌单"}
